// src/models/IoK8sApiAppsV1ControllerRevision/types.ts
import { IoK8sApimachineryPkgApisMetaV1ObjectMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ObjectMeta } from '../IoK8sApimachineryPkgApisMetaV1ObjectMeta';

export interface IoK8sApiAppsV1ControllerRevisionModelOptions {
  apiVersion?: string;
  data?: any;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ObjectMeta;
  revision: number;
}