// src/models/IoK8sApiAppsV1ControllerRevisionList/types.ts
import { IoK8sApiAppsV1ControllerRevisionModelOptions as IoK8sApiAppsV1ControllerRevision } from '../IoK8sApiAppsV1ControllerRevision';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiAppsV1ControllerRevisionListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiAppsV1ControllerRevision>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}